Bug Fix: Modal Rendering Issue
Problem: On first render Modal Data was not getting rendered correctly
Cause: The data flow and state management was designed considering that
modal will get remounted everytime when closed and reopened. Generally the Modal
is never unmounted and just closes. So the side effects were not getting executed as intended.

Solution: Instead of using Modal approach, Modal Like custom component is created
which takes cardDetails and its parent category as it inputs. Whenever user clicks on
card, this cardWizard component is inserted in dom and when closed it is removed from dom,
resulting in desired behaviour.