BugFix: close button of Modal out of visible area
Close button of modal was going out of the visible area on small devices,
when the content size increased.

Fix: maximum height of modal for small devices is set to small viewport height.

Critical Bug: Found a critical bug while trying to delete dates and checklist.
The firestore was throwing error as it expected the deleted fields to set null,
but was getting undefined instead.

Fix: the delete action function was modified to set the fields startDate, dueDate & checkList to null instead of undefined.